ftbzero

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: bperraud <bperraud@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/28 15:23:43 by bperraud          #+#    #+#             */
-/*   Updated: 2022/02/06 15:22:28 by bperraud         ###   ########.fr       */
+/*   Updated: 2022/02/06 16:05:37 by bperraud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,14 +19,14 @@
 # include <stdio.h>
 
 char	*get_next_line(int fd);
-char	*get_next_line_2(int ret, char *line, char *buff, char *temp);
-void	*ft_memchr(const void *s, int c, size_t n);
+char	*get_next_line_2(long ret, char *line, char *buff, char *temp);
+void	*ft_memchr(const void *s, int c, long n);
 char	*ft_strjoin(char const *s1, char const *s2);
-char    *ft_strncpy(char *dest, const char *src, unsigned int n);
+char    *ft_strncpy(char *dest, const char *src, long n);
 char	**ft_split(char const *s, const char c);
 size_t  ft_strlen(const char *s);
 char	*ft_strdup(const char *s1);
 char	*end_of_line(char *buff, char *line);
-void	ft_bzero(void *s, size_t n);
+void	ft_bzero(void *s, long n);
 
 #endif
